update user lobby info on joining room, set up a timer to delete lobby if lobby is empty to prevent unintentional deletion from interenet issue or page refresh

+++ api/socketServer.js
@@ -2,6 +2,7 @@ const prisma = require("../prisma")
 
 const roomPaths = {}
 const pendingClear = {}
+const pendingRoomDelete = {}
 
 const userCleanUp = async(socket) => {
   if (socket.hasCleaned) return;
@@ -16,12 +17,22 @@ const userCleanUp = async(socket) => {
       where: { lobbyId: Number(socket.roomId) },
     });
     if (players.length === 0) {
-      await prisma.lobby.delete({
-        where: { id: Number(socket.roomId) },
-      });
-      console.log(`Lobby ${socket.roomId} deleted due to being empty`);
-      delete roomPaths[socket.roomId];
-      delete pendingClear[socket.roomId];
+      if (pendingRoomDelete[socket.roomId]) return;
+      console.log(`Lobby ${socket.roomId} is empty. Starting deletion`)
+      const timeoutId = setTimeout(async() =>{
+        try {
+          await prisma.lobby.delete({
+            where: { id: Number(socket.roomId) },
+          });
+          console.log(`Lobby ${socket.roomId} deleted due to being empty`);
+          delete roomPaths[socket.roomId];
+          delete pendingClear[socket.roomId];
+        } catch (error) {
+          console.error("Error deleting empty room:", error)          
+        }
+        delete pendingRoomDelete[socket.roomId];
+      }, 10000);
+      pendingRoomDelete[socket.roomId] = timeoutId
     }
   } catch (error) {
     console.error("error cleaning up empty lobby:", error)
@@ -33,7 +44,7 @@ const setupSocket = (io) => {
   io.on("connection", (socket) => {
     console.log("a user connected", socket.id)
     
-    socket.on("joinRoom", ({roomId, username, userId}) => {
+    socket.on("joinRoom", async({roomId, username, userId}) => {
       if (!roomId || !username || !userId) {
         console.warn('Invalid joinRoom:', { roomId, username, userId });
         return;
@@ -42,6 +53,16 @@ const setupSocket = (io) => {
       socket.username = username
       socket.roomId = roomId
       socket.userId = userId
+      socket.hasCleaned = false
+      if (pendingRoomDelete[roomId]) {
+        clearTimeout(pendingRoomDelete[roomId])
+        delete pendingRoomDelete[roomId]
+        console.log(`Deletion cancelled: User ${username} joined room ${roomId}`)
+      };
+      await prisma.user.update({
+        where:{id:Number(userId)},
+        data:{lobbyId: Number(roomId)}
+      });
       if (!roomPaths[roomId]) roomPaths[roomId] = []
       socket.emit("init-paths", roomPaths[roomId]);
       console.log (`user ${username} has joined the room ${roomId}`);
